Menu callbacks need index param; 2.0.0-alpha.
The only reason that this commit leads to a 2.0.0-alpha release is
because it introduces backward-incompatible changes to the API (despite
it being alpha so it really doesn't matter anyway, but still), and thus
the major version must be incremented as per the "Semantic Versioning"
specification.

As a result, the alpha milestone is expected to have many major-version
updates, as is the nature of an alpha. Upon a 1.0.0-beta and a 1.0.0-rc
there will be much fewer major version updates that break the API.