Fonts each store a shader effect for rendering.
This change is put in place in order to facilitate differentiation of
shaders when they refer to different contexts. This will be implemented
using the `owner` field in the `asset::zAsset` class. Because multiple
windows can have fonts, they cannot refer to a static effect instance
that's tied to whichever context used a font first.